HOUSEKEEPING: Break websockethandler into websocket module
