+++ src/mozillavpn.h
@@ -32,7 +32,7 @@
 #include "statusicon.h"
 #include "telemetry.h"
 #include "theme.h"
-#include "websockethandler.h"
+#include "websocket/websockethandler.h"
 
 #include <QList>
 #include <QNetworkReply>
