+++ src/qmake/sources.pri
@@ -152,7 +152,8 @@ SOURCES += \
         update/versionapi.cpp \
         urlopener.cpp \
         update/webupdater.cpp \
-        websockethandler.cpp
+        websocket/exponentialbackoffstrategy.cpp \
+        websocket/websockethandler.cpp
 
 HEADERS += \
         addonmanager.h \
@@ -305,7 +306,8 @@ HEADERS += \
         update/versionapi.h \
         update/webupdater.h \
         urlopener.h \
-        websockethandler.h
+        websocket/exponentialbackoffstrategy.h \
+        websocket/websockethandler.h
 
 # Signal handling for unix platforms
 unix {
