+++ src/websocket/websockethandler.h
@@ -5,34 +5,14 @@
 #ifndef WEBSOCKETHANDLER_H
 #define WEBSOCKETHANDLER_H
 
+#include "exponentialbackoffstrategy.h"
+
 #include <QWebSocket>
 #include <QTimer>
 
-constexpr const int BASE_RETRY_INTERVAL_SEC = 5;               // 5s
-constexpr const int MAX_RETRY_INTERVAL_MSEC = 60 * 60 * 1000;  // 1hr
 constexpr const int WEBSOCKET_PING_INTERVAL_MSEC = 30 * 1000;  // 30s
 
-class ExponentialBackoffStrategy final : public QObject {
-  Q_OBJECT
- public:
-  ExponentialBackoffStrategy();
-  int scheduleNextAttempt();
-  void reset();
-
-#ifdef UNIT_TEST
-  void testOverrideBaseRetryInterval(int newInterval);
-  void testOverrideMaxRetryInterval(int newInterval);
-#endif
-
- signals:
-  void executeNextAttempt();
-
- private:
-  QTimer m_retryTimer;
-  int m_retryCounter = 1;
-  int m_maxInterval = MAX_RETRY_INTERVAL_MSEC;
-  int m_baseInterval = BASE_RETRY_INTERVAL_SEC;
-};
+class ExponentialBackoffStrategy;
 
 class WebSocketHandler final : public QObject {
   Q_OBJECT
